+++ docs/guides/changelog.md
@@ -11,6 +11,7 @@ description: |-
 
 - (Fix) Corrected URL encoding for multiple resources
 - (Fix) `fmc_device`: add missing FTDv100 performance tier
+- (Fix) `fmc_vpn_s2s_endpoints`: import
 - (Enhancement) Add support for Security Cloud Control (SCC) Firewall Management Base URI
 - (Enhancement) Add support for `fmc_sla_monitor` and `fmc_sla_monitors` resources and data sources
 - (Enhancement) Add support for SLA Monitors under IPv4 static routes